Set initial timestamp to the time the image was created
Previously, when docuum first starts, it would set the timestamp of every
image to the time docuum started, meaning old images would be removed at
random. Now, we use the time the image was created as the initial timestamp
(which is of course updated when that image is used like normal).