Add `--keep` argument to prevent deleting images
`--keep` is a new argument that allows users to specify one or more
regular expressions which, when matched against an image name, will
prevent that image from being deleted.

This is particularly useful for images that may require long re-build
times and so should not be cleaned up as part of normal system
maintenance.