Remove the dependency with Docker CLI.
Change the code to use the Docker API directly, so it doesn't
require additional binaries to exist in the container.

Change the Dockerfile to build the application statically,
allowing to build it with a Scratch image, making it distroless.
That way, you don't have to worry about updating versions of the base image,
because there is nothing else except the docuum binary.